+++ pkg/provider/provider.go
@@ -2,7 +2,6 @@ package provider
 
 import (
 	"context"
-	"time"
 )
 
 type ProviderType string
@@ -23,50 +22,6 @@ type ProviderSpec struct {
 type ProviderMetadata struct {
 	Name   string            `yaml:"name"`
 	Labels map[string]string `yaml:"labels"`
-}
-
-type ProviderConfig struct {
-	Transport TransportConfig `yaml:"transport"`
-	Payload   PayloadConfig   `yaml:"payload"`
-	Methods   []MethodConfig  `yaml:"methods"`
-}
-
-type TransportConfig struct {
-	Address string        `yaml:"address"`
-	Timeout time.Duration `yaml:"timeout"`
-	Logging LoggingConfig `yaml:"logging"`
-}
-
-type LoggingConfig struct {
-	Enabled bool `yaml:"enabled"`
-}
-
-type PayloadConfig struct {
-	Headers map[string]string `yaml:"headers"`
-}
-
-type MethodConfig struct {
-	Package  string         `yaml:"package"`
-	Service  string         `yaml:"service"`
-	Method   string         `yaml:"method"`
-	Type     ProviderType   `yaml:"type"`
-	Timeout  time.Duration  `yaml:"timeout"`
-	Filter   FilterConfig   `yaml:"filter"`
-	Request  RequestConfig  `yaml:"request"`
-	Response ResponseConfig `yaml:"response"`
-}
-
-type FilterConfig struct {
-	If string `yaml:"if"`
-}
-
-type RequestConfig struct {
-	Domain    string `yaml:"domain"`
-	DomainIDs string `yaml:"domain_ids"`
-}
-
-type ResponseConfig struct {
-	ItemID string `yaml:"itemId"`
 }
 
 type Provider interface {
